Expose evpl_add_poll and evpl_remove_poll in the public API
Move the evpl_poll type and callback declarations into a new public
header include/evpl/evpl_poll.h, included via evpl/evpl.h, and export
evpl_add_poll/evpl_remove_poll with SYMBOL_EXPORT so they can be linked
by consumers outside libevpl. The internal core/poll.h is reduced to
the struct definition and now pulls its typedefs from the public
header, matching the doorbell layering.

Adds an API reference page (docs/api/polls.md) and links it from the
API index.
