+++ include/evpl/evpl_poll.h
@@ -1,0 +1,36 @@
+// SPDX-FileCopyrightText: 2025 Ben Jarvis
+//
+// SPDX-License-Identifier: LGPL-2.1-only
+
+#pragma once
+
+#ifndef EVPL_INCLUDED
+#error "Do not include evpl_poll.h directly, include evpl/evpl.h instead"
+#endif /* ifndef EVPL_INCLUDED */
+
+struct evpl_poll;
+
+typedef void (*evpl_poll_enter_callback_t)(
+    struct evpl *evpl,
+    void        *private_data);
+
+typedef void (*evpl_poll_exit_callback_t)(
+    struct evpl *evpl,
+    void        *private_data);
+
+typedef void (*evpl_poll_callback_t)(
+    struct evpl *evpl,
+    void        *private_data);
+
+struct evpl_poll *
+evpl_add_poll(
+    struct evpl               *evpl,
+    evpl_poll_enter_callback_t enter_callback,
+    evpl_poll_exit_callback_t  exit_callback,
+    evpl_poll_callback_t       callback,
+    void                      *private_data);
+
+void
+evpl_remove_poll(
+    struct evpl      *evpl,
+    struct evpl_poll *poll);
