+++ src/core/poll.h
@@ -4,18 +4,8 @@
 
 #pragma once
 
-
-typedef void (*evpl_poll_enter_callback_t)(
-    struct evpl *evpl,
-    void        *private_data);
-
-typedef void (*evpl_poll_exit_callback_t)(
-    struct evpl *evpl,
-    void        *private_data);
-
-typedef void (*evpl_poll_callback_t)(
-    struct evpl *evpl,
-    void        *private_data);
+#define EVPL_INTERNAL 1
+#include "evpl/evpl.h"
 
 struct evpl_poll {
     evpl_poll_enter_callback_t enter_callback;
@@ -23,16 +13,3 @@ struct evpl_poll {
     evpl_poll_callback_t       callback;
     void                      *private_data;
 };
-
-struct evpl_poll *
-evpl_add_poll(
-    struct evpl               *evpl,
-    evpl_poll_enter_callback_t enter_callback,
-    evpl_poll_exit_callback_t  exit_callback,
-    evpl_poll_callback_t       callback,
-    void                      *private_data);
-
-void
-evpl_remove_poll(
-    struct evpl      *evpl,
-    struct evpl_poll *poll);
